// src/types/screens/planetsType.ts
export type PlanetasType = {
    clima: string
    creado: string
    diametro: string
    editado: string
    gravedad: string
    nombre: string
    orbital_periodo: string
    poblacion: string
    rotacion_periodo: string
    superficie_agua: string
    terrero: string
    url: string
    residentes: string[]
    peliculas: string[]
}

export type PlanetasInglesType = {
    climate: string
    created: string
    diameter: string
    edited: string
    gravity: string
    name: string
    orbital_period: string
    population: string
    rotation_period: string
    surface_water: string
    terrain: string
    url: string
    residents: string[]
    films: string[]
}